# Speedup new_with_metadata by removing sort